refactor: add roles to the entity with one for loop instead of two
Build functions params are objects instead of arrays.

Using for in loop create role property in every movie object which is array and fill it with role objects connected with this movie.
Using this approach, it is not necessary to iterate all movies with second loop.

Using the same functionality for actors