refactor: delete roles and actors' roles functionality on movie delete
On movie delete, get the roles which are connected with deleted movie and delete them. Extract only ids' of the actors who are connected with this roles than also remove deleted roles' ids from every actors' roles array.

With this approach .filter and .forEach will be executed on arrays with few elements.